Add a crosshair, controllable via cvars.
Add a crosshair, enabled by default, to address the complaint that it
is hard to aim when using the 3D weapon models.  Set the cvar
"crosshair_enabled" to 0 to disable.  The cvars
"crosshair_outer_radius", "crosshair_inner_radius",
"crosshair_thickness" and "crosshair_alpha" can be used to customise
its appearance:
* src/video.h, src/gl_draw.c: add function to draw the crosshair if
  enabled
* src/render.h, src/render.c: add cvars for crosshair
* README.md: note this change