Fix slow Fake Hitler fire projectiles when framerate is not low.
http://wolfenstein.wikia.com/wiki/NewWolf?oldid=19338 notes that "On
higher resolutions, the projectiles fired by Fake Hitlers travel much
more slowly than they are intended to."  As noted in
http://wolf3d.darkbb.com/t2089-bringing-newwolf-back-to-life#24350,
this is incorrect - the issue occurs with higher frame rates, not
higher resolutions, and occurred in the original game too.
http://maniacsvault.net/ecwolf/notes.php points out that the problem
is that, for these particular projectiles, the movement function is
called once per frame instead of once per tic.  This fix calls the
function from the correct slot.