Don't keep moving forward after holding down 'forward' for a while and using mouse.
If you walk or run forward for a while and use the mouse to turn while
doing this, after you release the button for 'forward', you may keep
moving forward for a while.  The amount of time that you keep moving
forward is increased based on the amount of time that 'forward' was
held, the amount of mouse movement used, and the lower the framerate
gets.  This issue was originally reported in
http://web.archive.org/web/20030608162220/http://hosting.gotdotnet.ru/DarkOne/phorum/showthread.asp?id=106

During debugging, it was found that the key handling code was stopping
the forward movement as soon as the WM_KEYUP event was received, but
reception of this event was delayed until some time after the key was
actually released.

The problem was that only a single Windows message was processed each
frame, so that as the framerate got lower and the amount of user input
increased, the backlog of unprocessed messages became quite large,
causing the WM_KEYUP event to not be processed until quite some time
after the key was released and the message was sent to the process.

This fix causes the message handling code to process all available
messages each time it checks for messages (once per frame).