Secret push walls should stop moving after at most 2 blocks, not 3.
As reported at http://forums.3drealms.com/vb/showthread.php?t=35359,
in episode 6 map 2, the when the secret push wall in the south-west
corner of the map moves 3 blocks, it prevents access to the rest of
the secret.  That thread references a bug in the original Wolfenstein
3D where push walls would sometimes move up to 3 blocks (taking into
consideration collisions), sometimes 2, being more likely to move up
to 3 blocks with higher frame rates, but that is not relevant in this
case.  The code in NewWolf for push walls is modified or rewritten,
and does not appear to suffer from this non-determinism.  However, it
does explicitly limit push walls to moving at most 3 blocks, which is
not consistent with the original Wolfenstein 3D code that attempted to
stop them after 2 blocks.

This fix changes the limit to 2 blocks for consistency with
Wolfenstein 3D and Wolf4SDL 1.7.