Briefly increase the crosshair size when picking up a bonus.
The NewWolf source included commented-out code to draw a crosshair.
This code referred to a function and a texture definition that do not
exist, but it was evident that the size of the crosshair was intended
to briefly increase when picking up a bonus.

This commit adds a similar feature to the crosshair code added in
NewWolfLives and removes the commented-out code.  The magnitude of the
increase in crosshair size can be controlled by the
"crosshair_bonus_scale" cvar.

It is hoped that the result would be approved by DarkOne.