Use CMake for generation of Microsoft Visual Studio solution/project files.
To make it easier to control build flags, and to potentially support
other versions of Visual Studio, other compilers and other platforms
in the future, this commit replaces the Visual Studio solution/project
files with a CMakeLists.txt file that can be used to generate them.
README.md is updated to describe how this generation is performed.

The CMakeLists.txt should introduce minimal changes to the source file
lists presented in the Visual Studio IDE and to the build flags used.
The most significant change is that the "Release" configuration is
replaced with "RelWithDebInfo", which is CMake's standard name for an
optimised release build with debugging information, which is what the
old "Release" configuration was.

Output directories are also quite different from those created by
newwolf.vcproj - they are the CMake defaults now - as putting the
executables in an "exe" directory outside of the rest of the build
area is not a good idea: it breaks the ability to have multiple build
directories (e.g. for different versions of Visual Studio) with one
source tree/working copy.  Mappings from old to new paths:
exe/NewWolfLives_d.exe -> <build directory>/debug/NewWolfLives_d.exe
exe/NewWolfLives.exe -> <build directory>/relwithdebinfo/NewWolfLives.exe
prj_vs80/Debug/*.obj -> <build directory>NewWolfLives.dir/Debug/*.obj
prj_vs80/Release/*.obj -> <build directory>NewWolfLives.dir/RelWithDebInfo/*.obj

Notes on some of the settings in newwolf.vcproj that are not included
in CMakeLists.txt:
* MinimalRebuild="true": this is documented as only affecting C++, so
  is not relevant
* TargetMachine="1": this setting forces X86, but the default setting,
  which is to automatically detect the target, appears to work
  properly, at least when run on 32-bit Windows.  It may be necessary
  to use a CMake setting to control the target platform if builds are
  performed on 64-bit Windows.
* @todo comments are included for settings that are missing compared
  to newwolf.vcproj but may be added in the future.

Note that the "OldHeaders" filter/group shown in the IDE is now
clearly marked as "(still used)" to address a comment in the WolfGL3D
sources pointing out that this was confusing.