Add support for Visual C++ 2008 and 2010, including Express Editions.
As it was found that the code can be built and run with Microsoft
Visual C++ 2008, Visual C++ 2008 Express Edition and Visual C++ 2010
Express Edition, this commit relaxes the version check in
CMakeLists.txt and updates the details on supported environments in
README.md.

A note is also added to CMakeLists.txt regarding a compiler option
that was recommended in a @todo comment but is deprecated in Visual
C++ 2008.  The 2010 documentation has not yet been checked to
determine if any other changes should be made to CMakeLists.txt.